requirements.txt
pandas
numpy
pyodbc
python-dotenv
scikit-learn
imbalanced-learn
seaborn
matplotlib
scipy
dill 
Flask
# -e .